Alter implementation to support more binding types
This version brings a large amount of support, including:
  - Configure via a wrangler config file or config object
  - Support KV, R2, DurableObjects, queues, static assets and services
  - Support mocking of services

There remains an issue in which if static assets are defined by the
configuration, the static asset server "steals" all routes and treats
them as files, which causes standard worker routes to report 404
errors (unless you hit an endpoint that's a static asset, in which case
you get the asset contents instead).

I'm unsure as to why this is currently happening, but since there is
probably not a lot of call for that in test suites at the moment, we
are setting that aside for now.